\name{Tushare-package}
\alias{Tushare-package}
\alias{Tushare}
\docType{package}
\title{
\packageTitle{Tushare}
}
\description{
Helps the R users to get data from "Tushare Pro".
"Tushare Pro" is a platform as well as a community with a lot of staffs working in financial area.
We support financial data such as stock price, financial report statements and digital coins data.
}
